pyproject.toml
[tool.poetry]
name = "codex"
version = "0.1.0"
description = "Personal Knowledge Store"
authors = ["khushmeeet <khushmeetsingh1996@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
libsass = "^0.21.0"
django-compressor = "^2.4.1"
django-sass-processor = "^1.0.1"
django-widget-tweaks = "^1.4.8"
django-libsass = "^0.8"
Markdown = "^3.3.4"
markdownify = "^0.9.0"
python-dotenv = "^0.18.0"
django-allow-cidr = "^0.3.1"
gunicorn = "^20.1.0"
djongo = "^1.3.6"
dnspython = "^2.1.0"
requests = "^2.25.1"
newspaper3k = "^0.2.8"
textract = "^1.6.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
